third_party/WebKit/LayoutTests/fast/dom/custom/element-upgrade-expected.txt

Tests the element upgrade algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS host.firstChild instanceof A is true
PASS element instanceof B is true
PASS callCount is 1
PASS element instanceof C is false
PASS Object.getPrototypeOf(element) is HTMLElement.prototype
PASS host.firstChild instanceof D is false
PASS document.querySelector(":unresolved") is host.firstChild
PASS upgradedOrder is ["e5","e4","e3","e2","e1"]
PASS successfullyParsed is true

TEST COMPLETE
